[jk] Reorganize date utility methods
